Inlcude testdata, tests, and modifications to runTestData for itertools.count
In order to test itertools.count, and any itertools functionality, itertools_
test.go needs to run test data that is added to testdata/itertools.star.

In order for count to be picked up in the test, runTestData needed to be
modified to include ItertoolsModule members. The method used in this commit,
i.e. updating Module.Members to include key/vals from ItertoolsModule.Members,
is definitely not the correct way to do this, since it beats the point of
having a separate module: overwrites of Module.Members from name collisions
can occur and the itertools namespace is left out.

Future commits or pull requests **must** resolve this issue.