Fix countObject initialization and reflect changes in test data
The previous iteration did not reflect the correct initialized count
object when one argument is passed in to the count function. When one
argument is passed in, said argument should define the start value and
the step value will default to one.

These changes are also tested with new tests in itertools.star.