Modify interface functions for floatOrInt to make use of interface type casting
The previous iteration used a struct that had pointers to starlark.Values with
a relatively complex and manual XOR relationship. This commit attempts to
rectify this by making use of type casting the underlying starlark.Value to
avoid potential pitfalls from using the above implementation.

The function `eq` on the receiver floatOrInt is also removed since a it was
only being used once in the String() method of countObject and a simple
check and computation sufficed.