Add hash functionality to countObject and incorporate tests
Cpython's count object's hash function, tp_hash, is inherited from
object. In some implementations of Python, object's hash is calculated
by right shifting id(object) by 4: id() >> 4. In this implementation
of countObject we simply return the id of a uuid, where the uuid package
used is github.com/google/uuid, that is stored on the countObject.

Tests for hash-ability of count object are included and replicate expected
Python output.