Use starlark.Binary for addition in next() instead of complex structure
The previous iteration used a complex adding method on a floatOrInt type receiver
in order to handle Float/Int combination arguments.

This method worked but added unnecessary complexity when compared with starlark's
Binary function.

Therefore, the previous adding method is removed, as well as the implementation of
floatOrInt (which was only used to support the addition of combo args for count and
step), and replaced in favor of the leaner starlark.Binary — alongside strict checks
for numeric types in the count_ function.